Add link[rel=atlernate] with feed url
This is the canonical way to indicate your site supports feeds. Some
browsers will show an RSS icon in the url bar when a website serves this
tag.

Also puts the feed_path into the configuration, to prevent duplicaiton.
Also moved the site_url into the config, such that when we switch to
HTTPS for example you only need to save it in one place.